+++ lib/helper/WebDriver.js
@@ -36,6 +36,7 @@ import { highlightElement } from './scripts/highlightElement.js'
 import { focusElement } from './scripts/focusElement.js'
 import { blurElement } from './scripts/blurElement.js'
 import { dontSeeElementError, seeElementError, seeElementInDOMError, dontSeeElementInDOMError } from './errors/ElementAssertion.js'
+import { dropFile } from './scripts/dropFile.js'
 import { dontSeeTraffic, seeTraffic, grabRecordedNetworkTraffics, stopRecordingTraffic, flushNetworkTraffics } from './network/actions.js'
 import WebElement from '../element/WebElement.js'
 
@@ -1376,20 +1377,12 @@ class WebDriver extends Helper {
     const targetRes = res.length ? res : await this._locate(locator)
     assertElementExists(targetRes, locator, 'Element')
     const targetEl = usingFirstElement(targetRes)
-    const base64Content = base64EncodeFile(file)
-    const fileName = path.basename(file)
-    const mimeType = getMimeType(fileName)
-    return this.browser.execute(function (el, data) {
-      var binaryStr = atob(data.base64Content)
-      var bytes = new Uint8Array(binaryStr.length)
-      for (var i = 0; i < binaryStr.length; i++) bytes[i] = binaryStr.charCodeAt(i)
-      var fileObj = new File([bytes], data.fileName, { type: data.mimeType })
-      var dataTransfer = new DataTransfer()
-      dataTransfer.items.add(fileObj)
-      el.dispatchEvent(new DragEvent('dragenter', { dataTransfer: dataTransfer, bubbles: true }))
-      el.dispatchEvent(new DragEvent('dragover', { dataTransfer: dataTransfer, bubbles: true }))
-      el.dispatchEvent(new DragEvent('drop', { dataTransfer: dataTransfer, bubbles: true }))
-    }, targetEl, { base64Content, fileName, mimeType })
+    const fileData = {
+      base64Content: base64EncodeFile(file),
+      fileName: path.basename(file),
+      mimeType: getMimeType(path.basename(file)),
+    }
+    return this.browser.execute(dropFile, targetEl, fileData)
   }
 
   /**
